Corrijo tipos de variables para qu se coincidan con la base de datos. -Borro datos de prueba hechos sobre "Main".

+++ ProductoVendido.cs
@@ -2,10 +2,10 @@
 {
     public class ProductoVendido
     {
-        private uint _id;
-        private uint _idProducto;
+        private long _id;
+        private long _idProducto;
         private int _stock;
-        private uint _idVenta;
+        private long _idVenta;
 
         public ProductoVendido()
         {
@@ -15,7 +15,7 @@
             this._idVenta = 0;
         }
 
-        public ProductoVendido(uint id, uint idProducto, int stock, uint idVenta)
+        public ProductoVendido(long id, long idProducto, int stock, long idVenta)
         {
             this._id = id;
             this._idProducto = idProducto;
@@ -23,9 +23,9 @@
             this._idVenta = idVenta;
         }
 
-        public uint Id { get => _id; set => _id = value; }
-        public uint IdProducto { get => _idProducto; set => _idProducto = value; }
+        public long Id { get => _id; set => _id = value; }
+        public long IdProducto { get => _idProducto; set => _idProducto = value; }
         public int Stock { get => _stock; set => _stock = value; }
-        public uint IdVenta { get => _idVenta; set => _idVenta = value; }
+        public long IdVenta { get => _idVenta; set => _idVenta = value; }
     }
 }
